PDI-1524 : Stream lookup breaks between 3.1.0-M2 and 3.1.0-RC1
git-svn-id: svn://source.pentaho.org/svnkettleroot/Kettle/trunk@8805 5fb7f6ec-07c1-534a-b4ca-9155e429e800

+++ src-core/org/pentaho/di/core/row/ValueMeta.java
@@ -1500,9 +1500,9 @@ public class ValueMeta implements ValueMetaInterface
         case TYPE_INTEGER:
             switch(storageType)
             {
-            case STORAGE_TYPE_NORMAL:         return BigDecimal.valueOf( ((Long)object).doubleValue() );
+            case STORAGE_TYPE_NORMAL:         return BigDecimal.valueOf( ((Long)object).longValue() );
             case STORAGE_TYPE_BINARY_STRING:  return BigDecimal.valueOf( ((Long)convertBinaryStringToNativeType((byte[])object)).longValue() );
-            case STORAGE_TYPE_INDEXED:        return BigDecimal.valueOf( ((Long)index[((Integer)object).intValue()]).doubleValue() );
+            case STORAGE_TYPE_INDEXED:        return BigDecimal.valueOf( ((Long)index[((Integer)object).intValue()]).longValue() );
             default: throw new KettleValueException(toString()+" : Unknown storage type "+storageType+" specified.");
             }
         case TYPE_NUMBER:
